Add structured skills-to-learn output in deep review and integrate into level-up pipeline
- Add `skills_to_learn` field to DeepReview interface and prompt
- Replace fragile gap text parsing with structured LLM output in
  level-up generation
- Add `/analyze` and `/resolve` API endpoints for skill analysis flow
- Redesign LevelUpPage with category filter, inline analysis, and
  resolved item management

+++ src/db/client.ts
@@ -509,10 +509,10 @@ export async function upsertLevelUpItem(
 ): Promise<void> {
   await pool.query(
     `INSERT INTO level_up_items (skill_name, category, source_job_ids)
-     VALUES ($1, $2, ARRAY[$3])
+     VALUES ($1, $2, ARRAY[$3::bigint])
      ON CONFLICT (skill_name) DO UPDATE
      SET source_job_ids = (
-       SELECT array_agg(DISTINCT x) FROM unnest(level_up_items.source_job_ids || ARRAY[$3]) AS x
+       SELECT array_agg(DISTINCT x) FROM unnest(level_up_items.source_job_ids || ARRAY[$3::bigint]) AS x
      ),
      updated_at = NOW()`,
     [skillName, category, sourceJobId]
